# Sample APIS/SampleInvetoryAPI/SampleInvetoryAPI/InvetoryAPIswagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "SampleInvetoryAPI",
        "version": "1.0"
    },
    "paths": {
        "/api/inventory": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryItem"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryItem"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/inventory/{id}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventory"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryItem"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryItem"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "InventoryItem": {
                "required": [
                    "productName",
                    "sku"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productName": {
                        "type": "string",
                        "nullable": true
                    },
                    "sku": {
                        "type": "string",
                        "nullable": true
                    },
                    "stockQuantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "warehouseLocation": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "additionalProperties": false
            }
        }
    }
}